:truck: Rename(mypage): 컴포넌트 위치 변경
Related to: #56
